+++ server/controllers/profileController.js
@@ -1,64 +1,61 @@
-const User = require("../modals/UserModel");
-const UserProfile = require("../modals/UserProfileModel");
+const AppDataSource = require("../config/database");
+const UserEntity = require("../entities/UserEntity");
+const UserProfileEntity = require("../entities/UserProfileEntity");
 const fs = require("fs");
 const path = require("path");
 
+const userRepository = AppDataSource.getRepository(UserEntity);
+const userProfileRepository = AppDataSource.getRepository(UserProfileEntity);
+
 const profileStaff = async (req, res) => {
   try {
-    const profile_id = req.params.id;
+    const profile_id = req.params.id; // User ID
     const { name, email, address, mobile, gender } = req.body;
     const newImage = req.file ? req.file.filename : null;
 
-    const userProfile = await UserProfile.findOne({ user_id: profile_id });
-    const user = await User.findById(profile_id);
+    const userProfile = await userProfileRepository.findOneBy({ user: { id: parseInt(profile_id) } });
+    const user = await userRepository.findOneBy({ id: parseInt(profile_id) });
 
     if (userProfile) {
-      // If a new image is provided, delete the old image
-      if (newImage) {
-        if (userProfile.image) {
-          const oldImagePath = path.join(
-            __dirname,
-            "../uploads/profile",
-            userProfile.image
-          );
-          fs.unlink(oldImagePath, (err) => {
-            if (err) console.error("Error deleting old image:", err);
-          });
-        }
+      if (newImage && userProfile.image) {
+        const oldImagePath = path.join(
+          __dirname,
+          "../uploads/profile",
+          userProfile.image
+        );
+        fs.unlink(oldImagePath, (err) => {
+          if (err) console.error("Error deleting old image:", err);
+        });
       }
 
-      // user model name and email update
       if (!user) {
-        res.status(404).send({ msg: "user not found" });
+        return res.status(404).send({ msg: "user not found" });
       }
 
-      const editProfile = await User.findByIdAndUpdate(
-        profile_id,
-        {
-          name: name || user.name,
-          email: email || user.email,
-        },
-        { new: true }
-      );
+      // Update User
+      user.name = name || user.name;
+      user.email = email || user.email;
+      await userRepository.save(user);
 
-      const updateProfile = await UserProfile.findByIdAndUpdate(
-        userProfile._id,
-        {
-          address: address || userProfile.address,
-          mobile: mobile || userProfile.mobile,
-          gender: gender || userProfile.gender,
-          image: newImage || userProfile.image,
-        },
-        { new: true }
-      );
+      // Update UserProfile
+      userProfile.address = address || userProfile.address;
+      userProfile.mobile = mobile || userProfile.mobile;
+      userProfile.gender = gender || userProfile.gender;
+      userProfile.image = newImage || userProfile.image;
+      await userProfileRepository.save(userProfile);
 
-      res.status(200).send({ profile: updateProfile, user: editProfile });
+      res.status(200).send({ profile: userProfile, user });
     } else {
-      res.status(404).send({ msg: "User not found" });
+      res.status(404).send({ msg: "User profile not found" });
     }
   } catch (error) {
+    console.error(error);
     res.status(500).send({ msg: error.message });
   }
+};
+
+module.exports = {
+  profileStaff,
 };
 
 module.exports = {
